Move currency to toolbar
