+++ client/packages/purchasing/src/purchase_order/DetailView/SidePanel/SupplierDetailSection.tsx
@@ -6,12 +6,7 @@ import {
   PanelRow,
   PanelLabel,
   NumericTextInput,
-  useNotification,
 } from '@openmsupply-client/common';
-import {
-  CurrencyAutocomplete,
-  CurrencyRowFragment,
-} from '@openmsupply-client/system';
 import { PurchaseOrderFragment } from '../../api';
 
 const slotProps = {
@@ -38,27 +33,10 @@ export const SupplierDetailSection = ({
   onChange,
 }: SupplierDetailSectionProps): ReactElement => {
   const t = useTranslation();
-  const { warning } = useNotification();
 
   const handleSupplierDiscountChange = (value: number | undefined) => {
     if (value == null || value === draft?.supplierDiscountPercentage) return;
     onChange({ supplierDiscountPercentage: value });
-  };
-
-  const handleCurrencyChange = (currency: CurrencyRowFragment | null) => {
-    if (!currency) return;
-    onChange({
-      currencyId: currency.id,
-      foreignExchangeRate: currency.rate,
-    });
-  };
-
-  const handleForeignExchangeRateChange = (value: number | undefined) => {
-    if (value == null || value === draft?.foreignExchangeRate) return;
-    if (draft?.foreignExchangeRate !== value) {
-      warning(t('warning.foreign-exchange-rate-different'))();
-      onChange({ foreignExchangeRate: value });
-    }
   };
 
   return (
@@ -80,30 +58,6 @@ export const SupplierDetailSection = ({
             max={100}
             disabled
             slotProps={slotProps}
-          />
-        </PanelRow>
-        <PanelRow>
-          <PanelLabel>{t('label.currency')}</PanelLabel>
-          <CurrencyAutocomplete
-            currencyId={draft?.currencyId}
-            onChange={handleCurrencyChange}
-            width={100}
-            sx={{
-              '& .MuiInputBase-root': {
-                backgroundColor: 'white',
-              },
-            }}
-            disabled={!!draft?.confirmedDatetime}
-          />
-        </PanelRow>
-        <PanelRow>
-          <PanelLabel>{t('label.foreign-exchange-rate')}</PanelLabel>
-          <NumericTextInput
-            value={draft?.foreignExchangeRate ?? 0}
-            onChange={handleForeignExchangeRateChange}
-            decimalLimit={4}
-            slotProps={slotProps}
-            disabled={!!draft?.confirmedDatetime}
           />
         </PanelRow>
       </Grid>
